use compiler-jx version of compc to build typedefs
It doesn't really matter either way, but we tend to use that version everywhere else, so it feels weird to use the "older" compiler. Plus, compiler-jx has certain defaults that are different for Royale, and it's probably better if those match up with the defaults that we use to build royale-asjs libraries. For instance, we made identifier names less strict for Royale specifically so that we could create typedefs using that feature.